Fjerner slettKontaktadresse.
