+++ src/main/kotlin/no/nav/tms/personopplysninger/api/personalia/personaliaRoute.kt
@@ -23,11 +23,6 @@ fun Route.personalia(personaliaService: HentPersonaliaService, oppdaterPersonali
         val resp = oppdaterPersonaliaService.slettTelefonNummer(call.user, telefonnummer)
         call.respond(resp)
     }
-
-    post("/slettKontaktadresse") {
-        val resp = oppdaterPersonaliaService.slettKontaktadresse(call.user)
-        call.respond(resp)
-    }
 }
 
 data class TelefonnummerEndring(
